correction requête + gestion medicament/categorie

+++ src/main/java/com/cours/api_gsbpharmastock/repository/CategoryRepository.java
@@ -3,5 +3,5 @@ package com.cours.api_gsbpharmastock.repository;
 import com.cours.api_gsbpharmastock.model.Category;
 import org.springframework.data.repository.CrudRepository;
 
-public interface CategoryRepository extends CrudRepository<Category, Long> {
+public interface CategoryRepository extends CrudRepository<Category, Integer> {
 }
